chore(lints): zero-warning workspace under clippy 1.98 — new-lint sweep
The container reset brought a fresh toolchain (rustc/clippy 1.98.1) whose
new lints flagged ~120 sites in pre-existing code (unnecessary_cast,
doc_lazy_continuation, needless_range_loop, is_multiple_of,
manual_range_contains, or_default...). All fixes are behavior-preserving
EXCEPT one: LightData's derived Default (empty sections Vec) was NOT
equivalent to LightData::new() (16 None slots) — clippy's or_default()
suggestion had silently broken 3 light differential tests; the manual
Default impl now delegates to new(), making or_default() safe.

Gates: cargo build/test/clippy --workspace --no-default-features all
green; 650 tests passing (628 baseline + 22 new UI Phase 1 tests, which
land in the next commit).

+++ voxelcraft/crates/vc-particles/src/particles.rs
@@ -91,6 +91,10 @@ impl ParticleSystem {
     /// vanilla break burst: a 4×4×4 grid of sub-tile particles across the
     /// block volume, launched with randomized velocities. `sky`/`blk` are
     /// the light levels at the block (baked brightness), `biome` its column.
+    // 8 params: the vanilla break burst carries position, block, biome
+    // and the two baked light levels — a struct here would just mirror
+    // the call site, so the lint is silenced deliberately.
+    #[allow(clippy::too_many_arguments)]
     pub fn spawn_block_break(
         &mut self,
         wx: i32,
@@ -101,7 +105,7 @@ impl ParticleSystem {
         sky: u8,
         blk: u8,
     ) {
-        let state = block as u16;
+        let state = block;
         let tile = state_tiles(state)[3]; // side tile like vanilla (top for grass tops)
         let tint_col = tint::block_tint_color(block, biome);
         let light = particle_light(sky, blk);
@@ -158,7 +162,7 @@ impl ParticleSystem {
         sky: u8,
         blk: u8,
     ) {
-        let tile = state_tiles(block as u16)[3];
+        let tile = state_tiles(block)[3];
         let tx = (tile % 32) as f32;
         let ty = (tile / 32) as f32;
         // vanilla Particles setting: Minimal/Decreased thin these out too
